fix bugs in ForkJoin crawler omplementation

+++ src/ru/fsl/CrawlManager.java
@@ -4,5 +4,5 @@ import java.net.URL;
 import java.util.Map;
 
 public interface CrawlManager {
-    Map<String, Integer> crawl(URL url, int maxDeepLevel);
+    Map<String, Long> crawl(URL url, int maxDeepLevel);
 }
